pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lacuna"
version = "0.1.0"
description = "Agentic C/C++ vulnerability scanner powered by Claude"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.25.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "docker>=7.0.0",
    "click>=8.1.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "types-PyYAML>=6.0.0",
]

[project.scripts]
lacuna = "lacuna.cli:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["lacuna*"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v"
